refactor: screen 분리 및 유효성 검사 로직 적용
